+++ BoBNolepAssique/Controllers/Intro/Kebutuhan/AddKebutuhanViewController.swift
@@ -7,23 +7,6 @@
 
 import UIKit
 
-struct SectionDetail {
-    let title: String
-    let options: [TableInputType]
-    let footer: String
-}
-
-enum TableInputType {
-    case switchCell(model: TableInputDetails)
-}
-struct TableInputDetails {
-    let title: String
-    let icon: UIImage?
-    let iconBackgrounColor: UIColor
-    var isOn: Bool
-    let handler: (() -> Void)
-}
-
 class AddKebutuhanViewController: ViewController {
     
     var models = [SectionDetail]()
